Various attempts to fix the "server hangs when printing A0" issue
I cannot get to the bottom of it - the issue is not reproducible on my
PC, including when running under Docker.

This commit does at least fix an issue where we weren't closing the
browser when an exception occurred. However, the key thing about our
A0 issue, is that no exception occurs.